fix: update deploy buttons and fix startup
