--- _data/db.json
@@ -1,12 +1,0 @@
-{
-  "appointments": [
-    {
-      "_id": "fksvcex249w",
-      "name": "Appointment2",
-      "phoneNumber": "+66666666",
-      "notification": "15",
-      "timeZone": "Africa/Algiers",
-      "time": "2016-02-17T08:00:00.000Z"
-    }
-  ]
-}
